Add a `RegisterGateway` function that we can use for JSON gateways.
This seemed to be the least messy way of adding this functionality without
effectively having to tear all of the server package apart.

I've had to pass more parameters into the method than I wanted, because
those I needed were not exposed. We also don't currently have the option
of securing this behind HTTPS, but I guess the lbproxy will handle that
for us.